+++ pages/_app.jsx
@@ -1,8 +1,9 @@
 import { createGlobalStyle, ThemeProvider } from 'styled-components';
 import { theme } from '../utils/ThemeConfig.jsx';
-import * as ga from '../lib/ga';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
+import Script from 'next/script';
+import * as gtag from '../lib/gtag.js';
 
 const GlobalStyle = createGlobalStyle`
 *, *::before, *::after {
@@ -41,23 +42,38 @@ const GlobalStyle = createGlobalStyle`
 
 function MyApp({ Component, pageProps }) {
 	const router = useRouter();
-
 	useEffect(() => {
 		const handleRouteChange = (url) => {
-			ga.pageview(url);
+			gtag.pageview(url);
 		};
-		//When the component is mounted, subscribe to router changes
-		//and log those page views
 		router.events.on('routeChangeComplete', handleRouteChange);
-
-		// If the component is unmounted, unsubscribe
-		// from the event with the `off` method
+		router.events.on('hashChangeComplete', handleRouteChange);
 		return () => {
 			router.events.off('routeChangeComplete', handleRouteChange);
+			router.events.off('hashChangeComplete', handleRouteChange);
 		};
 	}, [router.events]);
 	return (
 		<>
+			{/* Global Site Tag (gtag.js) - Google Analytics */}
+			<Script
+				strategy='afterInteractive'
+				src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_TRACKING_ID}`}
+			/>
+			<Script
+				id='gtag-init'
+				strategy='afterInteractive'
+				dangerouslySetInnerHTML={{
+					__html: `
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', '${gtag.GA_TRACKING_ID}', {
+              page_path: window.location.pathname,
+            });
+          `,
+				}}
+			/>
 			<GlobalStyle />
 			<ThemeProvider theme={theme}>
 				<Component {...pageProps} />
