+++ pages/_document.jsx
@@ -1,4 +1,4 @@
-import Document, { Head } from 'next/document';
+import Document from 'next/document';
 import { ServerStyleSheet } from 'styled-components';
 
 export default class MyDocument extends Document {
@@ -10,38 +10,7 @@ export default class MyDocument extends Document {
 			ctx.renderPage = () =>
 				originalRenderPage({
 					enhanceApp: (App) => (props) =>
-						sheet.collectStyles(
-							<App {...props}>
-								<Head>
-									<link rel='preconnect' href='https://fonts.googleapis.com' />
-									<link
-										rel='preconnect'
-										href='https://fonts.gstatic.com'
-										crossOrigin
-									/>
-									<link
-										href='https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,400;0,500;0,700;1,700&family=Poppins:wght@500;600;700&display=swap'
-										rel='stylesheet'
-									/>
-									<script
-										async
-										src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}`}
-									/>
-									<script
-										dangerouslySetInnerHTML={{
-											__html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS}', {
-              page_path: window.location.pathname,
-            });
-          `,
-										}}
-									/>
-								</Head>
-							</App>
-						),
+						sheet.collectStyles(<App {...props}></App>),
 				});
 
 			const initialProps = await Document.getInitialProps(ctx);
